chore(WinUI3): fix typos and formatting
