+++ src/gui4/windows/kDrive client/kDrive client/CustomControls/QuickAccess.xaml.cs
@@ -11,7 +11,7 @@ namespace Infomaniak.kDrive.CustomControls
 {
     public sealed partial class QuickAccess : UserControl
     {
-        private AppModel _viewModel = (App.Current as App).Data;
+        private AppModel _viewModel = App.ServiceProvider.GetRequiredService<AppModel>();
         public AppModel ViewModel => _viewModel;
 
         public QuickAccess()
