+++ src/gui4/windows/kDrive client/kDrive client/Pages/Onboarding/OAuthLoadingPage.xaml.cs
@@ -19,11 +19,11 @@ namespace Infomaniak.kDrive.Pages.Onboarding
     public sealed partial class OAuthLoadingPage : Page, IDisposable
     {
         private static TimeSpan _oauthTimeOut = TimeSpan.FromMinutes(5);
-        private readonly AppModel _viewModel = (App.Current as App).Data;
+        private readonly AppModel _viewModel = App.ServiceProvider.GetRequiredService<AppModel>();
         private ViewModels.Onboarding? _onboardingViewModel;
 
         private Task? _onBoardingTask;
-        private CancellationTokenSource _OAuthCts = new(_oauthTimeOut);
+        private CancellationTokenSource _oauthCts = new(_oauthTimeOut);
 
         private Task? _enableRestartTask;
         private CancellationTokenSource _enableRestartCts = new();
@@ -46,7 +46,7 @@ namespace Infomaniak.kDrive.Pages.Onboarding
                 _onboardingViewModel = obvm;
                 _onboardingViewModel.PropertyChanged += OnOnboardingViewModelPropertyChanged;
 
-                _onBoardingTask = _onboardingViewModel.ConnectUser(_OAuthCts.Token);
+                _onBoardingTask = _onboardingViewModel.ConnectUser(_oauthCts.Token);
                 ScheduleRestartButtonEnableAsync();
             }
             else
@@ -132,7 +132,7 @@ namespace Infomaniak.kDrive.Pages.Onboarding
             try
             {
                 // Cancel any ongoing authentication attempt
-                await _OAuthCts.CancelAsync();
+                await _oauthCts.CancelAsync();
                 if (_onBoardingTask is not null)
                 {
                     await _onBoardingTask;
@@ -144,9 +144,9 @@ namespace Infomaniak.kDrive.Pages.Onboarding
             }
             finally
             {
-                _OAuthCts.Dispose();
-                _OAuthCts = new CancellationTokenSource(_oauthTimeOut);
-                _onBoardingTask = _onboardingViewModel?.ConnectUser(_OAuthCts.Token);
+                _oauthCts.Dispose();
+                _oauthCts = new CancellationTokenSource(_oauthTimeOut);
+                _onBoardingTask = _onboardingViewModel?.ConnectUser(_oauthCts.Token);
                 ScheduleRestartButtonEnableAsync();
             }
         }
@@ -200,8 +200,8 @@ namespace Infomaniak.kDrive.Pages.Onboarding
                 _onboardingViewModel.PropertyChanged -= OnOnboardingViewModelPropertyChanged;
             }
 
-            _OAuthCts.Cancel();
-            _OAuthCts.Dispose();
+            _oauthCts.Cancel();
+            _oauthCts.Dispose();
 
             _enableRestartCts.Cancel();
             _enableRestartCts.Dispose();
