+++ src/gui4/windows/kDrive client/kDrive client/Pages/Onboarding/WelcomePage.xaml.cs
@@ -24,7 +24,7 @@ namespace Infomaniak.kDrive.Pages.Onboarding
 {
     public sealed partial class WelcomePage : Page
     {
-        private AppModel _viewModel = (App.Current as App).Data;
+        private AppModel _viewModel = App.ServiceProvider.GetRequiredService<AppModel>();
         private ViewModels.Onboarding? _onBoardingViewModel;
         public AppModel ViewModel { get { return _viewModel; } }
         public WelcomePage()
