+++ src/gui4/windows/kDrive client/kDrive client/Pages/SettingsPage.xaml.cs
@@ -40,7 +40,7 @@ namespace Infomaniak.kDrive.Pages
 {
     public sealed partial class SettingsPage : Page
     {
-        private AppModel _viewModel = (App.Current as App).Data;
+        private AppModel _viewModel = App.ServiceProvider.GetRequiredService<AppModel>();
         public AppModel ViewModel => _viewModel;
         public SettingsPage()
         {
