+++ src/server/comm/socketcommserver.cpp
@@ -121,10 +121,10 @@ void SocketCommChannel::callbackHandler() {
 
 uint64_t SocketCommChannel::bytesAvailable() const {
     try {
-        return (std::max)(0, _socket.available());
+        return static_cast<uint64_t>((std::max)(0, _socket.available()));
     } catch (Poco::IOException &ex) {
         LOG_ERROR(Log::instance()->getLogger(), "Socket available error: " << ex.displayText());
-        return 0;
+        return static_cast<uint64_t>(0);
     }
 }
